# api/swagger.yaml
definitions:
  model.MessageDataTransferObject:
    properties:
      created_at:
        type: string
      id:
        type: string
      text:
        type: string
      updated_at:
        type: string
    type: object
  status.ErrorResponse:
    properties:
      code:
        type: integer
      details:
        items:
          $ref: '#/definitions/status.ErrorResponseDetail'
        type: array
      message:
        type: string
    type: object
  status.ErrorResponseDetail:
    properties:
      message:
        type: string
    type: object
info:
  contact: {}
paths:
  /v1/messages:
    get:
      description: Returns a list of all messages
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.MessageDataTransferObject'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/status.ErrorResponse'
      summary: Returns a list of all messages
      tags:
      - messages
    post:
      consumes:
      - application/json
      description: Creates a new message
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.MessageDataTransferObject'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/status.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/status.ErrorResponse'
      summary: Creates a new message
      tags:
      - messages
  /v1/messages/{id}:
    delete:
      description: Deletes an existing message by its id
      parameters:
      - description: Message ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/status.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/status.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/status.ErrorResponse'
      summary: Deletes an existing message by its id
      tags:
      - messages
    put:
      consumes:
      - application/json
      description: Updates an existing message by its id
      parameters:
      - description: Message ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.MessageDataTransferObject'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/status.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/status.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/status.ErrorResponse'
      summary: Updates an existing message by its id
      tags:
      - messages
swagger: "2.0"